chore: Move feature flag annotation and enum to interfaces
